Add experimental Dockerfile for end-to-end testing (WIP)
The new Dockerfile is based on .devcontainer/Dockerfile, and simplified
to allow optimzation for layer caching. Most notably, it
- introduces multi-stage builds, and
- copies sources from the build context, when they are used to
  invalidate caches, only when relevant files are updated

The goal is to run the Docker build in a GitHub Action (e.g.
https://github.com/docker/build-push-action), which requires strategic
caching to not exceed allowed build minutes and cache sizes.

**Caveat**
This is an experimental/explorative patch and does not fit in with the
rest of the toolchain. It:
* ignores bazel for a lighter toolchain and to focus on one
  caching mechanism
* moves all glibc-related build instructions into gen_sysroot.sh to
  reduce complexity

Depending on how well this works, we may replace or revert back to
existing tools.