ci: fix e2e cache re-use between branches
Prior to this PR e2e builds were only triggered und thus cached on PR.
This allows re-using caches for new pushes to a PR, but not for new PRs
from sibling branches. This is because GitHub only shares caches between
branches, if one branch is based off another.

As a fix, we also trigger and cache on push to main, so that PRs based
on main can re-use that cache.

fixes #256

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>