build: add release stage to Dockerfile.e2e
In #270 Dockerfile.e2e is optimized for faster build times and smaller
cache sizes. As a consequence, the resulting Docker image lacks parts of
the toolchain, and is not useful beyond running end-to-end tests on
`docker build`.

This commit adds an alternative 'release' stage to Dockerfile.e2e, which
produces a Docker image that does include the full toolchain, e.g. to run
demos, tests, experiments, etc.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>