build: move clang installation in Dockerfile.e2e
In preparation for upcoming build and test tool changes, this commit
moves the clang installation from the 'build-glibc' to the 'base' stage
in Dockerfile.e2e and adds the clang 'bin' directory to the PATH.

This allows us to use clang from PATH in 'release' and 'test'
end-stages, without copying it from the intermediate-'build-glibc'
stage, where it was installed previously.

In addition, the patch moves copying clang builtins, which are included
in the source code repository, from make_glibc_and_sysroot.sh, to the
'base' stage in Dockerfile.e2e.

Both, clang and the builtins, are not maintained by lind-wasm, so it
makes sense to treat them like other build dependencies, and not as part
of the build process.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>