build: replacement for lindtool.sh (#276)
* build: move clang installation in Dockerfile.e2e

In preparation for upcoming build and test tool changes, this commit
moves the clang installation from the 'build-glibc' to the 'base' stage
in Dockerfile.e2e and adds the clang 'bin' directory to the PATH.

This allows us to use clang from PATH in 'release' and 'test'
end-stages, without copying it from the intermediate-'build-glibc'
stage, where it was installed previously.

In addition, the patch moves copying clang builtins, which are included
in the source code repository, from make_glibc_and_sysroot.sh, to the
'base' stage in Dockerfile.e2e.

Both, clang and the builtins, are not maintained by lind-wasm, so it
makes sense to treat them like other build dependencies, and not as part
of the build process.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>

* build: add lind_compile and lind_run scripts

Replacement for `lindtool.sh {compile_test, run}`

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>

* test: replace lindtool.sh in wasmtestreport.py

Use lind_compile and lind_run scripts instead.

IMPORTANT: backwards-incompatible default usage!!

Previously, lindtool.sh
- `clang` was expected to be in
  `clang+llvm-16.0.4-x86_64-linux-gnu-ubuntu-22.04/bin/clang`.
- `wasmtime` was expected to be in `src/wasmtime/target/debug/wasmtime`

Now, via lind_compile and lind_run
- `clang` is expected to be on PATH
- `wasmtime` is expected to be on `src/wasmtime/target/release/wasmtime`

Make sure to add clang to the path and compile wasmtime with `--release`
fag, when using wasmtestreport.py (see Dockerfile.e2e for blueprint).

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>

* build: add Makefile and use in Dockerfile.e2e

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>

* build: fix comment

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>

* build: add cache invalidation warning to Dockerfile

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>

---------

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>