docs: minor polish (#290)
* docs: move toolchain and debug guides to contribute

Those docs should primarily be intended for contributors.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>

* docs: move toc from contribute overview page

As the contribute section evolves, it might make sense to remove the TOC
from the overview page, to not require manual syncing with the actual TOC.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>

* docs: move getting started page to top-level nav

With the recent move of other usage sub pages, the getting started doc
becomes the last page in the usage section, making the usage section
obsolete.

This patch moves it to the top-level nav, making it more prominent, and
removes the emptied usage section.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>

---------

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>